Payment API

+++ vendingmachine/src/main/java/za/co/tfradebe/vendingmachine/inventory/db/entities/ProductEntity.java
@@ -14,10 +14,10 @@ public class ProductEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private double price;
+    private int price;
     private int quantity;
 
-    public ProductEntity(String name, double price, int quantity) {
+    public ProductEntity(String name, int price, int quantity) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
